blog-py/blog/requirements.txt

beautifulsoup4==4.6.3
Django==2.1.4
django-cors-headers==2.4.0
django-filter==2.0.0
djangorestframework==3.9.0
Markdown==3.0.1
mysqlclient==1.3.14
Pillow==5.3.0
pytz==2018.7
